backend/requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.8
django-cors-headers==4.6.0
djangorestframework==3.15.2
google-genai==1.14.0
pydantic==2.12.5
python-dotenv==1.1.0
sqlparse==0.5.3
psycopg[binary]
